Make score a float
Make score a float for more precision

+++ Eval.h
@@ -1,5 +1,5 @@
 #ifndef _Eval_H_Included_
 #define _Eval_H_Included_
 
-int Evaluate_Position();
+float Evaluate_Position();
 #endif
